analysis/generate_supplementary_figures.py
# analysis/generate_supplementary_figures.py
"""
Generate supplementary figures S3-S6 for the TMBMC paper.
- S3: Constellation diagrams
- S4: SNR analysis (types I, II, III)
- S5: Confusion matrices
- S6: Energy per bit analysis
"""

import sys
from pathlib import Path
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns   # type: ignore
from typing import Dict, List
import yaml

# Add project root to path
project_root = Path(__file__).parent.parent
sys.path.append(str(project_root))

from src.config_utils import preprocess_config


def plot_figure_s3_constellation(results_dir: Path, save_path: Path):
    """
    Generate Figure S3: Constellation diagrams for all modulation schemes.
    Shows decision space (q_GLU vs q_GABA) with transmitted symbols marked.
    """
    fig, axes = plt.subplots(1, 3, figsize=(15, 5))
    modes = ['MoSK', 'CSK', 'Hybrid']
    
    for idx, mode in enumerate(modes):
        ax = axes[idx]
        
        # Try to load the SER data which contains statistics
        data_file = results_dir / "data" / f"ser_vs_nm_{mode.lower()}.csv"
        if not data_file.exists():
            ax.text(0.5, 0.5, f"No data for {mode}", 
                   ha='center', va='center', transform=ax.transAxes)
            ax.set_title(f"{mode} Constellation")
            continue
        
        # For demonstration, create synthetic constellation data
        # In real implementation, load from saved statistics
        np.random.seed(42)
        
        if mode == 'MoSK':
            # Binary: GLU (0) vs GABA (1)
            q_glu_0 = np.random.normal(1.0, 0.2, 100)  # When GLU sent
            q_gaba_0 = np.random.normal(0.0, 0.2, 100)
            q_glu_1 = np.random.normal(0.0, 0.2, 100)  # When GABA sent
            q_gaba_1 = np.random.normal(1.0, 0.2, 100)
            
            ax.scatter(q_glu_0, q_gaba_0, c='blue', alpha=0.5, label='Symbol 0 (GLU)', s=30)
            ax.scatter(q_glu_1, q_gaba_1, c='red', alpha=0.5, label='Symbol 1 (GABA)', s=30)
            
            # Decision boundary
            x = np.linspace(-0.5, 1.5, 100)
            ax.plot(x, x, 'k--', alpha=0.5, label='Decision boundary')
            
        elif mode == 'CSK':
            # 4-level on GLU channel
            for level in range(4):
                q_glu = np.random.normal(level/3, 0.15, 50)
                q_gaba = np.random.normal(0.0, 0.1, 50)
                ax.scatter(q_glu, q_gaba, alpha=0.5, label=f'Level {level}', s=30)
            
        else:  # Hybrid
            # 4 symbols: 00, 01, 10, 11
            symbols = [(0, 0), (0, 1), (1, 0), (1, 1)]
            colors = ['blue', 'cyan', 'red', 'orange']
            for (mol, amp), color in zip(symbols, colors):
                if mol == 0:  # GLU
                    q_glu = np.random.normal(0.5 + 0.5*amp, 0.15, 50)
                    q_gaba = np.random.normal(0.0, 0.1, 50)
                else:  # GABA
                    q_glu = np.random.normal(0.0, 0.1, 50)
                    q_gaba = np.random.normal(0.5 + 0.5*amp, 0.15, 50)
                ax.scatter(q_glu, q_gaba, c=color, alpha=0.5, 
                          label=f'Symbol {mol}{amp}', s=30)
        
        ax.set_xlabel('$q_{GLU}$ (nC)')
        ax.set_ylabel('$q_{GABA}$ (nC)')
        ax.set_title(f'({chr(97+idx)}) {mode} Constellation')
        ax.legend(fontsize=8)
        ax.grid(True, alpha=0.3)
    
    plt.suptitle('Figure S3: Constellation Diagrams', fontsize=14)
    plt.tight_layout()
    plt.savefig(save_path, dpi=300, bbox_inches='tight')
    plt.close()


def plot_figure_s4_snr_types(results_dir: Path, save_path: Path):
    """
    Generate Figure S4: SNR analysis showing Types I, II, and III.
    """
    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 5))
    
    # Synthetic data for demonstration
    distances = np.array([25, 50, 75, 100, 150, 200, 250])
    
    # Type I: Molecular (concentration-based)
    snr_i = 30 * np.exp(-distances / 100)  # dB, decays with distance
    
    # Type II: Electrical (after transduction)
    eta = 0.7  # Transduction efficiency
    snr_ii = snr_i - 10*np.log10(1/eta)
    
    # Type III: Digital (after detection)
    coding_gain = 3  # dB
    snr_iii = snr_ii + coding_gain
    
    # Panel (a): SNR vs Distance
    ax1.plot(distances, snr_i, 'b-o', label='Type I (Molecular)', linewidth=2)
    ax1.plot(distances, snr_ii, 'g--s', label='Type II (Electrical)', linewidth=2)
    ax1.plot(distances, snr_iii, 'r-.^', label='Type III (Digital)', linewidth=2)
    
    ax1.set_xlabel('Distance (μm)')
    ax1.set_ylabel('SNR (dB)')
    ax1.set_title('(a) SNR Types vs Distance')
    ax1.grid(True, alpha=0.3)
    ax1.legend()
    ax1.set_ylim([0, 35])
    
    # Panel (b): SNR Conversion Factors
    categories = ['Molecular\n→Electrical', 'Electrical\n→Digital']
    factors = [-10*np.log10(1/eta), coding_gain]
    colors = ['green', 'red']
    
    bars = ax2.bar(categories, factors, color=colors, alpha=0.7)
    ax2.set_ylabel('SNR Change (dB)')
    ax2.set_title('(b) SNR Conversion Factors')
    ax2.grid(True, axis='y', alpha=0.3)
    ax2.axhline(y=0, color='k', linestyle='-', linewidth=0.5)
    
    # Add value labels on bars
    for bar, val in zip(bars, factors):
        height = bar.get_height()
        ax2.text(bar.get_x() + bar.get_width()/2., height + 0.5*np.sign(height),
                f'{val:.1f} dB', ha='center', va='bottom' if val > 0 else 'top')
    
    plt.suptitle('Figure S4: Signal-to-Noise Ratio Analysis', fontsize=14)
    plt.tight_layout()
    plt.savefig(save_path, dpi=300, bbox_inches='tight')
    plt.close()


def plot_figure_s5_confusion_matrices(results_dir: Path, save_path: Path):
    """
    Generate Figure S5: Confusion matrices for all modulation schemes.
    """
    fig, axes = plt.subplots(1, 3, figsize=(15, 4))
    
    # Synthetic confusion matrices for demonstration
    # In real implementation, compute from actual results
    
    # MoSK (2x2)
    cm_mosk = np.array([[95, 5], [3, 97]]) / 100
    
    # CSK (4x4)
    cm_csk = np.array([
        [90, 8, 2, 0],
        [7, 85, 7, 1],
        [2, 6, 87, 5],
        [0, 1, 4, 95]
    ]) / 100
    
    # Hybrid (4x4)
    cm_hybrid = np.array([
        [92, 5, 2, 1],
        [4, 91, 1, 4],
        [3, 1, 93, 3],
        [1, 3, 4, 92]
    ]) / 100
    
    matrices = [cm_mosk, cm_csk, cm_hybrid]
    titles = ['MoSK', 'CSK-4', 'Hybrid']
    
    for idx, (cm, title) in enumerate(zip(matrices, titles)):
        ax = axes[idx]
        
        # Create heatmap
        im = ax.imshow(cm, cmap='Blues', vmin=0, vmax=1)
        
        # Add colorbar
        cbar = plt.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
        cbar.set_label('Probability', rotation=270, labelpad=15)
        
        # Set ticks
        n_classes = cm.shape[0]
        ax.set_xticks(np.arange(n_classes))
        ax.set_yticks(np.arange(n_classes))
        
        if n_classes == 2:
            labels = ['GLU', 'GABA']
        else:
            labels = [f'S{i}' for i in range(n_classes)]
        
        ax.set_xticklabels(labels)
        ax.set_yticklabels(labels)
        
        # Add text annotations
        for i in range(n_classes):
            for j in range(n_classes):
                text = ax.text(j, i, f'{cm[i, j]:.2f}',
                             ha="center", va="center",
                             color="white" if cm[i, j] > 0.5 else "black")
        
        ax.set_xlabel('Detected Symbol')
        ax.set_ylabel('Transmitted Symbol')
        ax.set_title(f'({chr(97+idx)}) {title} Confusion Matrix')
    
    plt.suptitle('Figure S5: Detection Confusion Matrices', fontsize=14)
    plt.tight_layout()
    plt.savefig(save_path, dpi=300, bbox_inches='tight')
    plt.close()


def plot_figure_s6_energy_per_bit(results_dir: Path, save_path: Path):
    """
    Generate Figure S6: Energy per bit analysis including transduction efficiency.
    """
    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 5))
    
    # Load config for parameters
    with open(project_root / "config" / "default.yaml") as f:
        config = yaml.safe_load(f)
    cfg = preprocess_config(config)
    
    # Parameters
    k_B = 1.381e-23  # Boltzmann constant (J/K)
    T = cfg['temperature_K']  # Temperature (K)
    eta = 0.1  # Transduction efficiency (10%)
    
    # Data
    Nm_values = np.logspace(2, 5, 50)
    
    # Calculate energy per bit for each scheme
    # MoSK: 1 bit per symbol
    E_mosk = Nm_values * k_B * T * np.log(2) / eta
    
    # CSK-4: 2 bits per symbol, but average Nm is lower
    E_csk = 0.5 * Nm_values * k_B * T * np.log(2) * 2 / eta
    
    # Hybrid: 2 bits per symbol, average energy between MoSK and CSK
    E_hybrid = 0.75 * Nm_values * k_B * T * np.log(2) * 2 / eta
    
    # Panel (a): Energy per bit vs Nm
    ax1.loglog(Nm_values, E_mosk * 1e15, 'b-', label='MoSK', linewidth=2)
    ax1.loglog(Nm_values, E_csk * 1e15, 'g--', label='CSK-4', linewidth=2)
    ax1.loglog(Nm_values, E_hybrid * 1e15, 'r-.', label='Hybrid', linewidth=2)
    
    # Landauer limit
    E_landauer = k_B * T * np.log(2)
    ax1.axhline(y=E_landauer * 1e15, color='k', linestyle=':', 
                label='Landauer limit', alpha=0.5)
    
    ax1.set_xlabel('Number of Molecules per Symbol (Nm)')
    ax1.set_ylabel('Energy per Bit (fJ)')
    ax1.set_title(f'(a) Energy Efficiency (η = {eta:.0%})')
    ax1.grid(True, alpha=0.3)
    ax1.legend()
    ax1.set_ylim([1e-3, 1e3])
    
    # Panel (b): Efficiency factor analysis
    eta_values = np.logspace(-3, 0, 50)
    Nm_fixed = 1e4
    
    E_vs_eta = Nm_fixed * k_B * T * np.log(2) / eta_values
    
    ax2.loglog(eta_values * 100, E_vs_eta * 1e15, 'k-', linewidth=2)
    ax2.axvline(x=eta*100, color='r', linestyle='--', alpha=0.5, 
                label=f'Current η = {eta:.0%}')
    
    ax2.set_xlabel('Transduction Efficiency η (%)')
    ax2.set_ylabel('Energy per Bit (fJ)')
    ax2.set_title(f'(b) Impact of Transduction Efficiency (Nm = {Nm_fixed:.0e})')
    ax2.grid(True, alpha=0.3)
    ax2.legend()
    
    # Add caption info
    caption = (f"Parameters: T = {T} K, k_B = {k_B:.3e} J/K, "
              f"η = {eta:.0%} (transduction efficiency)")
    fig.text(0.5, -0.02, caption, ha='center', fontsize=9, style='italic')
    
    plt.suptitle('Figure S6: Energy per Bit Analysis', fontsize=14)
    plt.tight_layout()
    plt.savefig(save_path, dpi=300, bbox_inches='tight')
    plt.close()


def main():
    """Generate all supplementary figures."""
    print("\n" + "="*60)
    print("Generating Supplementary Figures")
    print("="*60)
    
    # Setup paths
    results_dir = project_root / "results"
    figures_dir = results_dir / "figures"
    figures_dir.mkdir(parents=True, exist_ok=True)
    
    # Generate each figure
    print("\nGenerating Figure S3: Constellation diagrams...")
    plot_figure_s3_constellation(results_dir, figures_dir / "figS3_constellation.png")
    print("✓ Figure S3 saved")
    
    print("\nGenerating Figure S4: SNR analysis...")
    plot_figure_s4_snr_types(results_dir, figures_dir / "figS4_snr_types.png")
    print("✓ Figure S4 saved")
    
    print("\nGenerating Figure S5: Confusion matrices...")
    plot_figure_s5_confusion_matrices(results_dir, figures_dir / "figS5_confusion.png")
    print("✓ Figure S5 saved")
    
    print("\nGenerating Figure S6: Energy per bit...")
    plot_figure_s6_energy_per_bit(results_dir, figures_dir / "figS6_energy.png")
    print("✓ Figure S6 saved")
    
    print("\n" + "="*60)
    print("All supplementary figures generated successfully!")
    print(f"Results saved in: {figures_dir}")
    print("="*60)


if __name__ == "__main__":
    main()