pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tri-channel-oect-mc"
version = "0.1.0"
description = "Tri-channel OECT molecular communication simulator"
requires-python = ">=3.10"
dependencies = [
    "numpy==1.26.4",
    "scipy==1.13.0",
    "pandas==2.2.1",
    "matplotlib==3.8.3",
    "tqdm==4.66.2",
    "pyyaml==6.0.1",
    "pyarrow==15.0.1",
    "pytest==8.1.1",
    "jupyter==1.0.0",
]

[project.optional-dependencies]
dev = [
    "black==24.2.0",
    "mypy==1.9.0",
    "pytest-cov==4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
